+++ CHANGES.rst
@@ -8,6 +8,8 @@ Unreleased
 -   Use modern packaging metadata with ``pyproject.toml`` instead of ``setup.cfg``.
     :pr:`1793`
 -   Use ``flit_core`` instead of ``setuptools`` as build backend.
+-   Fixed `autoescape=False` not working for builtin filter `tojson`.
+    :issue:`1934`
 
 
 Version 3.1.3